% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_response.R
\name{AutoNLSBayesFit}
\alias{AutoNLSBayesFit}
\title{AutoNLS Bayesian fitted response object}
\description{
AutoNLS Bayesian fitted response object

AutoNLS Bayesian fitted response object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AutoNLSBayesFit-new}{\code{AutoNLSBayesFit$new()}}
\item \href{#method-AutoNLSBayesFit-fit}{\code{AutoNLSBayesFit$fit()}}
\item \href{#method-AutoNLSBayesFit-predict}{\code{AutoNLSBayesFit$predict()}}
\item \href{#method-AutoNLSBayesFit-posterior_curve}{\code{AutoNLSBayesFit$posterior_curve()}}
\item \href{#method-AutoNLSBayesFit-posterior_function}{\code{AutoNLSBayesFit$posterior_function()}}
\item \href{#method-AutoNLSBayesFit-control_coefficients}{\code{AutoNLSBayesFit$control_coefficients()}}
\item \href{#method-AutoNLSBayesFit-prior_table}{\code{AutoNLSBayesFit$prior_table()}}
\item \href{#method-AutoNLSBayesFit-derivative}{\code{AutoNLSBayesFit$derivative()}}
\item \href{#method-AutoNLSBayesFit-elasticity}{\code{AutoNLSBayesFit$elasticity()}}
\item \href{#method-AutoNLSBayesFit-incremental_response}{\code{AutoNLSBayesFit$incremental_response()}}
\item \href{#method-AutoNLSBayesFit-diagnostics}{\code{AutoNLSBayesFit$diagnostics()}}
\item \href{#method-AutoNLSBayesFit-draws}{\code{AutoNLSBayesFit$draws()}}
\item \href{#method-AutoNLSBayesFit-artifact}{\code{AutoNLSBayesFit$artifact()}}
\item \href{#method-AutoNLSBayesFit-clone}{\code{AutoNLSBayesFit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSBayesFit-new"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSBayesFit-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSBayesFit$new(
  data,
  x,
  y,
  family,
  controls,
  model_domain,
  priors,
  chains,
  iter,
  burnin,
  thin,
  seed,
  proposal_scale,
  control_standardize
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSBayesFit-fit"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSBayesFit-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSBayesFit$fit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSBayesFit-predict"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSBayesFit-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSBayesFit$predict(
  new_data,
  type = c("summary", "draws"),
  include_residual = FALSE,
  probs = c(0.025, 0.5, 0.975),
  control_draws = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSBayesFit-posterior_curve"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSBayesFit-posterior_curve}{}}}
\subsection{Method \code{posterior_curve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSBayesFit$posterior_curve(grid, controls, probs = c(0.025, 0.5, 0.975))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSBayesFit-posterior_function"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSBayesFit-posterior_function}{}}}
\subsection{Method \code{posterior_function()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSBayesFit$posterior_function(
  grid,
  type = c("summary", "draws"),
  probs = c(0.025, 0.5, 0.975)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSBayesFit-control_coefficients"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSBayesFit-control_coefficients}{}}}
\subsection{Method \code{control_coefficients()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSBayesFit$control_coefficients(probs = c(0.025, 0.5, 0.975))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSBayesFit-prior_table"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSBayesFit-prior_table}{}}}
\subsection{Method \code{prior_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSBayesFit$prior_table()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSBayesFit-derivative"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSBayesFit-derivative}{}}}
\subsection{Method \code{derivative()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSBayesFit$derivative(new_data, probs = c(0.025, 0.5, 0.975))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSBayesFit-elasticity"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSBayesFit-elasticity}{}}}
\subsection{Method \code{elasticity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSBayesFit$elasticity(new_data, probs = c(0.025, 0.5, 0.975))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSBayesFit-incremental_response"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSBayesFit-incremental_response}{}}}
\subsection{Method \code{incremental_response()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSBayesFit$incremental_response(x0, x1, controls)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSBayesFit-diagnostics"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSBayesFit-diagnostics}{}}}
\subsection{Method \code{diagnostics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSBayesFit$diagnostics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSBayesFit-draws"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSBayesFit-draws}{}}}
\subsection{Method \code{draws()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSBayesFit$draws()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSBayesFit-artifact"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSBayesFit-artifact}{}}}
\subsection{Method \code{artifact()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSBayesFit$artifact()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSBayesFit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSBayesFit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSBayesFit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
