% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autonls_vnext.R
\name{AutoNLSFit}
\alias{AutoNLSFit}
\title{AutoNLS vNext fitted result object}
\description{
AutoNLS vNext fitted result object

AutoNLS vNext fitted result object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AutoNLSFit-new}{\code{AutoNLSFit$new()}}
\item \href{#method-AutoNLSFit-fit}{\code{AutoNLSFit$fit()}}
\item \href{#method-AutoNLSFit-summary}{\code{AutoNLSFit$summary()}}
\item \href{#method-AutoNLSFit-metrics}{\code{AutoNLSFit$metrics()}}
\item \href{#method-AutoNLSFit-fitted_values}{\code{AutoNLSFit$fitted_values()}}
\item \href{#method-AutoNLSFit-residuals}{\code{AutoNLSFit$residuals()}}
\item \href{#method-AutoNLSFit-predict}{\code{AutoNLSFit$predict()}}
\item \href{#method-AutoNLSFit-score}{\code{AutoNLSFit$score()}}
\item \href{#method-AutoNLSFit-derivative}{\code{AutoNLSFit$derivative()}}
\item \href{#method-AutoNLSFit-elasticity}{\code{AutoNLSFit$elasticity()}}
\item \href{#method-AutoNLSFit-plots}{\code{AutoNLSFit$plots()}}
\item \href{#method-AutoNLSFit-compare_plot}{\code{AutoNLSFit$compare_plot()}}
\item \href{#method-AutoNLSFit-best_model}{\code{AutoNLSFit$best_model()}}
\item \href{#method-AutoNLSFit-artifacts}{\code{AutoNLSFit$artifacts()}}
\item \href{#method-AutoNLSFit-report}{\code{AutoNLSFit$report()}}
\item \href{#method-AutoNLSFit-clone}{\code{AutoNLSFit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSFit-new"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSFit-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSFit$new(
  data,
  x,
  y,
  models,
  weights_col,
  loss,
  n_starts,
  seed,
  optimizer,
  quantile_level,
  huber_delta,
  lower_bounds,
  upper_bounds,
  maxit,
  reltol,
  scale_x,
  scale_y,
  model_status,
  validation_fraction,
  validation_seed,
  interval_method,
  interval_level,
  interval_n,
  interval_seed,
  interval_models,
  interval_max_rows,
  start_strategy,
  theme
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSFit-fit"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSFit-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSFit$fit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSFit-summary"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSFit-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSFit$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSFit-metrics"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSFit-metrics}{}}}
\subsection{Method \code{metrics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSFit$metrics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSFit-fitted_values"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSFit-fitted_values}{}}}
\subsection{Method \code{fitted_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSFit$fitted_values(model = "best")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSFit-residuals"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSFit-residuals}{}}}
\subsection{Method \code{residuals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSFit$residuals(model = "best")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSFit-predict"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSFit-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSFit$predict(new_data, model = "best")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSFit-score"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSFit-score}{}}}
\subsection{Method \code{score()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSFit$score(new_data, model = "best")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSFit-derivative"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSFit-derivative}{}}}
\subsection{Method \code{derivative()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSFit$derivative(new_data, model = "best")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSFit-elasticity"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSFit-elasticity}{}}}
\subsection{Method \code{elasticity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSFit$elasticity(new_data, model = "best")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSFit-plots"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSFit-plots}{}}}
\subsection{Method \code{plots()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSFit$plots()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSFit-compare_plot"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSFit-compare_plot}{}}}
\subsection{Method \code{compare_plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSFit$compare_plot(grid = NULL, model = "all")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSFit-best_model"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSFit-best_model}{}}}
\subsection{Method \code{best_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSFit$best_model()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSFit-artifacts"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSFit-artifacts}{}}}
\subsection{Method \code{artifacts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSFit$artifacts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSFit-report"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSFit-report}{}}}
\subsection{Method \code{report()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSFit$report()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AutoNLSFit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AutoNLSFit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AutoNLSFit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
