% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_nls_methods.R
\name{BIC.custom_nls}
\alias{BIC.custom_nls}
\title{BIC for custom_nls objects}
\usage{
\method{BIC}{custom_nls}(object, ...)
}
\arguments{
\item{object}{A \code{custom_nls} object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
The BIC value for the model.
}
\description{
Computes the Bayesian Information Criterion (BIC) for models of class \code{custom_nls}.
}
\examples{
# Assuming `fit` is a fitted `custom_nls` model
BIC(fit)
}
