% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_nls_methods.R
\name{fitted.custom_nls}
\alias{fitted.custom_nls}
\title{Extract Fitted Values from a Custom NLS Object}
\usage{
\method{fitted}{custom_nls}(object, ...)
}
\arguments{
\item{object}{An object of class \code{custom_nls}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A numeric vector of fitted values.
}
\description{
Extracts the fitted values from a model of class \code{custom_nls}.
}
