% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts_vnext.R
\name{generate_autonls_artifacts}
\alias{generate_autonls_artifacts}
\title{Generate AutoNLS vNext artifacts}
\usage{
generate_autonls_artifacts(fit)
}
\arguments{
\item{fit}{An AutoNLSFit object.}
}
\value{
A named list of data.table artifacts.
}
\description{
Generate AutoNLS vNext artifacts
}
