% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_nls_methods.R
\name{logLik.custom_nls}
\alias{logLik.custom_nls}
\title{Calculate Log-Likelihood for a Custom NLS Object}
\usage{
\method{logLik}{custom_nls}(object, ...)
}
\arguments{
\item{object}{An object of class \code{custom_nls}.}

\item{...}{Additional arguments (ignored).}
}
\value{
The log-likelihood value as a numeric scalar.
}
\description{
Computes the log-likelihood for a model of class \code{custom_nls}.
}
