% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_registry.R
\name{nls_model_registry}
\alias{nls_model_registry}
\title{Built-in AutoNLS vNext model registry}
\usage{
nls_model_registry()
}
\value{
A named list of nonlinear model specifications.
}
\description{
Built-in AutoNLS vNext model registry
}
