% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{NonLinearFitter}
\alias{NonLinearFitter}
\title{Fit nonlinear regression models}
\description{
Fit nonlinear regression models

Fit nonlinear regression models
}
\details{
NonLinearFitter

An R6 class for automatically fitting non-linear regression models.
Includes a library of pre-defined models to simplify selection.

The choice of fitting method depends on the arguments:
\itemize{
\item If \code{force_optim = TRUE}, the function will use \code{optim()} regardless of whether weights are supplied.
\item If \code{weights} is supplied, \code{optim()} will be used even if \code{force_optim = FALSE}.
\item If neither \code{force_optim} nor \code{weights} is supplied, the function defaults to \code{nls()} for unweighted fitting.
}

\strong{Behavior Examples}:
\enumerate{
\item \strong{Default behavior}: \code{nls()} is used when \code{weights = NULL} and \code{force_optim = FALSE}.
\item \strong{Weighted fitting}: \code{optim()} is used when \code{weights} is provided, even if \code{force_optim = FALSE}.
\item \strong{Forced optimization}: \code{optim()} is used when \code{force_optim = TRUE}, regardless of whether \code{weights} is supplied.
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{A data.table containing the dataset for modeling.}

\item{\code{models}}{A list of non-linear models to test.}

\item{\code{fit_results}}{A list to store the results of model fits.}

\item{\code{evaluation_metrics}}{A list to store evaluation metrics for each model.}

\item{\code{plots}}{A list to store plots of model fits.}

\item{\code{model_library}}{A pre-defined library of common non-linear models.
Initialize the NonLinearFitter class}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NonLinearFitter-new}{\code{NonLinearFitter$new()}}
\item \href{#method-NonLinearFitter-list_models}{\code{NonLinearFitter$list_models()}}
\item \href{#method-NonLinearFitter-add_model}{\code{NonLinearFitter$add_model()}}
\item \href{#method-NonLinearFitter-fit_models}{\code{NonLinearFitter$fit_models()}}
\item \href{#method-NonLinearFitter-model_comparison_plot}{\code{NonLinearFitter$model_comparison_plot()}}
\item \href{#method-NonLinearFitter-clone}{\code{NonLinearFitter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonLinearFitter-new"></a>}}
\if{latex}{\out{\hypertarget{method-NonLinearFitter-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonLinearFitter$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A data.table containing the dataset for modeling.
Must include the predictor and response variable columns.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the NonLinearFitter class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonLinearFitter-list_models"></a>}}
\if{latex}{\out{\hypertarget{method-NonLinearFitter-list_models}{}}}
\subsection{Method \code{list_models()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonLinearFitter$list_models()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.table summarizing available models.
Add a non-linear model for testing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonLinearFitter-add_model"></a>}}
\if{latex}{\out{\hypertarget{method-NonLinearFitter-add_model}{}}}
\subsection{Method \code{add_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonLinearFitter$add_model(name, formula = NULL, start_params = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the model (e.g., "Hill").}

\item{\code{formula}}{The non-linear formula for the model (optional if using pre-defined model).}

\item{\code{start_params}}{A list of starting parameters for the model (optional if using pre-defined model).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonLinearFitter-fit_models"></a>}}
\if{latex}{\out{\hypertarget{method-NonLinearFitter-fit_models}{}}}
\subsection{Method \code{fit_models()}}{
\code{fit_models()} first standardizes your data before fitting the model.
The fitting method depends on whether a \code{weights_col} is provided: \code{nls()} is used
for unweighted fitting, while \code{optim()} is used for weighted fitting. The returned
parameters are based on the standardized data. However, when scoring models, the
results are back-transformed to align with the original data scale.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonLinearFitter$fit_models(
  x_col,
  y_col,
  weights_col = NULL,
  control = list(maxiter = 1024),
  force_optim = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x_col}}{The name of the predictor variable.}

\item{\code{y_col}}{The name of the response variable.}

\item{\code{weights_col}}{The name of the weights variable.}

\item{\code{control}}{A list of control parameters for the optimizer, such as \code{maxiter}.
Default is \code{list(maxiter = 200)}.}

\item{\code{force_optim}}{Logical; if \code{TRUE}, forces the use of \code{optim()} regardless of whether weights are supplied. Defaults to \code{FALSE}.}

\item{\code{...}}{Additional arguments to be passed to the underlying fitting functions
(\code{nlsLM} for unweighted models or \code{optim} for weighted models). Examples include
\code{trace}, \code{lower}, and \code{upper} for \code{nlsLM}, or \code{reltol}, \code{parscale}, and others for \code{optim}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of fitted model objects.
Generate a comparison plot for model shapes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonLinearFitter-model_comparison_plot"></a>}}
\if{latex}{\out{\hypertarget{method-NonLinearFitter-model_comparison_plot}{}}}
\subsection{Method \code{model_comparison_plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonLinearFitter$model_comparison_plot(
  x_range = seq(1, 100, by = 1),
  normalize = TRUE,
  theme = "macarons"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x_range}}{A numeric vector specifying the range of x values to evaluate.}

\item{\code{normalize}}{Logical. If TRUE, normalizes the y values for each model to fall between 0 and 1.
Defaults to TRUE.}

\item{\code{theme}}{A string specifying the plot theme (e.g., "macarons").}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{echarts4r} object representing the comparison plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonLinearFitter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NonLinearFitter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonLinearFitter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
