% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_scoring.R
\name{NonLinearModelScorer}
\alias{NonLinearModelScorer}
\title{NonLinearModelScorer}
\description{
An R6 class to score non-linear regression models on new data
and visualize the results.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fit_results}}{A list of fitted model objects.}

\item{\code{scored_data}}{A list of data.tables containing scored data.}

\item{\code{score_plots}}{A list of plots visualizing scored data.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NonLinearModelScorer-new}{\code{NonLinearModelScorer$new()}}
\item \href{#method-NonLinearModelScorer-score_new_data}{\code{NonLinearModelScorer$score_new_data()}}
\item \href{#method-NonLinearModelScorer-generate_score_plot}{\code{NonLinearModelScorer$generate_score_plot()}}
\item \href{#method-NonLinearModelScorer-clone}{\code{NonLinearModelScorer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonLinearModelScorer-new"></a>}}
\if{latex}{\out{\hypertarget{method-NonLinearModelScorer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonLinearModelScorer$new(fit_results)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit_results}}{A list of fitted model objects (e.g., output from NonLinearFitter).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the NonLinearModelScorer class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonLinearModelScorer-score_new_data"></a>}}
\if{latex}{\out{\hypertarget{method-NonLinearModelScorer-score_new_data}{}}}
\subsection{Method \code{score_new_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonLinearModelScorer$score_new_data(new_data, x_col)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_data}}{A data.table containing the new data to score.}

\item{\code{x_col}}{The predictor column in \code{new_data}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of data.tables with predicted values for each model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonLinearModelScorer-generate_score_plot"></a>}}
\if{latex}{\out{\hypertarget{method-NonLinearModelScorer-generate_score_plot}{}}}
\subsection{Method \code{generate_score_plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonLinearModelScorer$generate_score_plot(
  model_name,
  new_data,
  x_col,
  theme = "westeros"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model_name}}{The name of the model to plot.}

\item{\code{new_data}}{The original new data used for scoring.}

\item{\code{x_col}}{The predictor column in \code{new_data}.}

\item{\code{theme}}{Echarts theme}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A plot visualizing the scored data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonLinearModelScorer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NonLinearModelScorer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonLinearModelScorer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
