% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_nls_methods.R
\name{predict.custom_nls}
\alias{predict.custom_nls}
\title{Predict Values from a Custom NLS Object}
\usage{
\method{predict}{custom_nls}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class \code{custom_nls}.}

\item{newdata}{A \code{data.frame} or \code{data.table} containing the predictor variable \code{x}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A numeric vector of predicted values.
}
\description{
Uses a fitted \code{custom_nls} model to predict values for new data.
}
