% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts_vnext.R
\name{qa_autonls_curve_artifact_contract}
\alias{qa_autonls_curve_artifact_contract}
\title{QA checks for AutoQuant-consumable curve artifact contract}
\usage{
qa_autonls_curve_artifact_contract()
}
\value{
TRUE invisibly when checks pass.
}
\description{
QA checks for AutoQuant-consumable curve artifact contract
}
