% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts_vnext.R
\name{qa_autonls_interval_estimation}
\alias{qa_autonls_interval_estimation}
\title{QA checks interval estimation contract}
\usage{
qa_autonls_interval_estimation()
}
\value{
TRUE invisibly when checks pass.
}
\description{
QA checks interval estimation contract
}
