% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts_vnext.R
\name{qa_autonls_intervals}
\alias{qa_autonls_intervals}
\title{QA checks for optional interval estimation}
\usage{
qa_autonls_intervals()
}
\value{
TRUE invisibly when checks pass.
}
\description{
QA checks for optional interval estimation
}
