% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts_vnext.R
\name{qa_autonls_model_confidence}
\alias{qa_autonls_model_confidence}
\title{QA checks for model confidence}
\usage{
qa_autonls_model_confidence()
}
\value{
TRUE invisibly when checks pass.
}
\description{
QA checks for model confidence
}
