% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{R6Class}
\alias{e_chart}
\alias{gam}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{echarts4r}{\code{\link[echarts4r:init]{e_chart}}}

  \item{mgcv}{\code{\link[mgcv]{gam}}}

  \item{R6}{\code{\link[R6]{R6Class}}}
}}

