% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_nls_methods.R
\name{residuals.custom_nls}
\alias{residuals.custom_nls}
\title{Extract Residuals from a Custom NLS Object}
\usage{
\method{residuals}{custom_nls}(object, ...)
}
\arguments{
\item{object}{An object of class \code{custom_nls}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A numeric vector of residuals.
}
\description{
Extracts the residuals of a fitted model of class \code{custom_nls}.
}
