% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_nls_methods.R
\name{summary.custom_nls}
\alias{summary.custom_nls}
\title{Summary for a Custom NLS Object}
\usage{
\method{summary}{custom_nls}(object, ...)
}
\arguments{
\item{object}{An object of class \code{custom_nls}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A list containing model coefficients, residuals, and fitted values.
}
\description{
Provides a summary of the fitted model of class \code{custom_nls}.
}
