% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_nls_methods.R
\name{AIC.custom_nls}
\alias{AIC.custom_nls}
\title{AIC for custom_nls objects}
\usage{
\method{AIC}{custom_nls}(object, ...)
}
\arguments{
\item{object}{A \code{custom_nls} object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
The AIC value for the model
}
\description{
Computes the Akaike Information Criterion for models of class \code{custom_nls}.
}
