% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_scoring.R
\name{ModelScorer}
\alias{ModelScorer}
\title{ModelScorer}
\description{
An R6 class to score non-linear regression models on new data
and visualize the results.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelScorer-new}{\code{ModelScorer$new()}}
\item \href{#method-ModelScorer-score_new_data}{\code{ModelScorer$score_new_data()}}
\item \href{#method-ModelScorer-generate_score_plot}{\code{ModelScorer$generate_score_plot()}}
\item \href{#method-ModelScorer-clone}{\code{ModelScorer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelScorer-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelScorer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelScorer$new(fit_results)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelScorer-score_new_data"></a>}}
\if{latex}{\out{\hypertarget{method-ModelScorer-score_new_data}{}}}
\subsection{Method \code{score_new_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelScorer$score_new_data(
  new_data,
  x_col,
  get_prediction_bounds = FALSE,
  lower_bound = 0.025,
  upper_bound = 0.975,
  n_sim = 1000,
  allow_missing_features = FALSE,
  return_cols = NULL,
  id_col = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelScorer-generate_score_plot"></a>}}
\if{latex}{\out{\hypertarget{method-ModelScorer-generate_score_plot}{}}}
\subsection{Method \code{generate_score_plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelScorer$generate_score_plot(
  model_name,
  x_col,
  theme = "westeros",
  title = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelScorer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelScorer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelScorer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
