% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{NonLinearFitter}
\alias{NonLinearFitter}
\title{NonLinearFitter}
\description{
An R6 class for automatically fitting non-linear regression models.
Includes a library of pre-defined models to simplify selection.
}
\examples{

## ------------------------------------------------
## Method `NonLinearFitter$new`
## ------------------------------------------------

data <- data.table::data.table(x = 1:100, y = 5 / (1 + exp(-0.1 * (1:100 - 50))))
fitter <- NonLinearFitter$new(data)

## ------------------------------------------------
## Method `NonLinearFitter$add_model`
## ------------------------------------------------

# Add a pre-defined model
fitter$add_model("Hill")

# Add a custom model
fitter$add_model("Custom", y ~ a * exp(-b * x), list(a = 1, b = 0.1))

## ------------------------------------------------
## Method `NonLinearFitter$fit_models`
## ------------------------------------------------

fitter$fit_models(x_col = "x", y_col = "y", control = list(maxiter = 200))
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{A data.table containing the dataset for modeling.}

\item{\code{models}}{A list of non-linear models to test.}

\item{\code{fit_results}}{A list to store the results of model fits.}

\item{\code{evaluation_metrics}}{A list to store evaluation metrics for each model.}

\item{\code{plots}}{A list to store plots of model fits.}

\item{\code{model_library}}{A pre-defined library of common non-linear models.
Initialize the NonLinearFitter class}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NonLinearFitter-new}{\code{NonLinearFitter$new()}}
\item \href{#method-NonLinearFitter-list_models}{\code{NonLinearFitter$list_models()}}
\item \href{#method-NonLinearFitter-add_model}{\code{NonLinearFitter$add_model()}}
\item \href{#method-NonLinearFitter-fit_models}{\code{NonLinearFitter$fit_models()}}
\item \href{#method-NonLinearFitter-clone}{\code{NonLinearFitter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonLinearFitter-new"></a>}}
\if{latex}{\out{\hypertarget{method-NonLinearFitter-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonLinearFitter$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A data.table containing the dataset for modeling.
Must include the predictor and response variable columns.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the NonLinearFitter class.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{data <- data.table::data.table(x = 1:100, y = 5 / (1 + exp(-0.1 * (1:100 - 50))))
fitter <- NonLinearFitter$new(data)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonLinearFitter-list_models"></a>}}
\if{latex}{\out{\hypertarget{method-NonLinearFitter-list_models}{}}}
\subsection{Method \code{list_models()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonLinearFitter$list_models()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.table summarizing available models.
Add a non-linear model for testing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonLinearFitter-add_model"></a>}}
\if{latex}{\out{\hypertarget{method-NonLinearFitter-add_model}{}}}
\subsection{Method \code{add_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonLinearFitter$add_model(name, formula = NULL, start_params = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the model (e.g., "Hill").}

\item{\code{formula}}{The non-linear formula for the model (optional if using pre-defined model).}

\item{\code{start_params}}{A list of starting parameters for the model (optional if using pre-defined model).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Add a pre-defined model
fitter$add_model("Hill")

# Add a custom model
fitter$add_model("Custom", y ~ a * exp(-b * x), list(a = 1, b = 0.1))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonLinearFitter-fit_models"></a>}}
\if{latex}{\out{\hypertarget{method-NonLinearFitter-fit_models}{}}}
\subsection{Method \code{fit_models()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonLinearFitter$fit_models(x_col, y_col, control = list(maxiter = 200))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x_col}}{The name of the predictor variable.}

\item{\code{y_col}}{The name of the response variable.}

\item{\code{control}}{A list of control parameters for the optimizer, such as \code{maxiter}.
Default is \code{list(maxiter = 200)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of fitted model objects.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fitter$fit_models(x_col = "x", y_col = "y", control = list(maxiter = 200))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonLinearFitter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NonLinearFitter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonLinearFitter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
