% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{NonLinearModelEvaluator}
\alias{NonLinearModelEvaluator}
\title{NonLinearModelEvaluator}
\description{
An R6 class to evaluate non-linear regression models.
Includes tools to generate tables of statistics and visualizations
to compare models against data.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fit_results}}{A list of fitted model objects.}

\item{\code{evaluation_metrics}}{A data.table containing model evaluation metrics.}

\item{\code{plots}}{A list of visualizations comparing models against data.}

\item{\code{data}}{The original dataset used for fitting models.
Initialize the NonLinearModelEvaluator class}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NonLinearModelEvaluator-new}{\code{NonLinearModelEvaluator$new()}}
\item \href{#method-NonLinearModelEvaluator-generate_metrics}{\code{NonLinearModelEvaluator$generate_metrics()}}
\item \href{#method-NonLinearModelEvaluator-generate_comparison_plot}{\code{NonLinearModelEvaluator$generate_comparison_plot()}}
\item \href{#method-NonLinearModelEvaluator-clone}{\code{NonLinearModelEvaluator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonLinearModelEvaluator-new"></a>}}
\if{latex}{\out{\hypertarget{method-NonLinearModelEvaluator-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonLinearModelEvaluator$new(fit_results, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fit_results}}{A list of fitted model objects (e.g., output from NonLinearFitter).}

\item{\code{data}}{The original dataset used for fitting models.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the NonLinearModelEvaluator class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonLinearModelEvaluator-generate_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-NonLinearModelEvaluator-generate_metrics}{}}}
\subsection{Method \code{generate_metrics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonLinearModelEvaluator$generate_metrics()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data.table of evaluation metrics.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonLinearModelEvaluator-generate_comparison_plot"></a>}}
\if{latex}{\out{\hypertarget{method-NonLinearModelEvaluator-generate_comparison_plot}{}}}
\subsection{Method \code{generate_comparison_plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonLinearModelEvaluator$generate_comparison_plot(data, x_col, y_col)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A \code{data.table} or \code{data.frame} containing the dataset used for evaluation.}

\item{\code{x_col}}{A string specifying the name of the x variable in the dataset.}

\item{\code{y_col}}{A string specifying the name of the y variable in the dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{echarts4r} plot showing observed vs. predicted data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonLinearModelEvaluator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NonLinearModelEvaluator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonLinearModelEvaluator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
